requirements.txt
bunq_sdk 
openai
requests
discord